completed base authorization modules

+++ routes/web.php
@@ -15,33 +15,33 @@ Route::middleware(['auth', 'verified'])->group(function () {
 		return Inertia::render('dashboard');
 	})->name('dashboard');
 
-	Route::get('/permissions', [PermissionController::class, 'index'])->name('permissions.index');
-	Route::post('/permissions', [PermissionController::class, 'store'])->name('permissions.store');
-	Route::put('/permissions/{permission}', [PermissionController::class, 'update'])->name('permissions.update');
-	Route::delete('/permissions/{permission}', [PermissionController::class, 'destroy'])->name('permissions.destroy');
+	Route::get('/permissions', [PermissionController::class, 'index'])->name('permissions.index')->can('view_permissions');
+	Route::post('/permissions', [PermissionController::class, 'store'])->name('permissions.store')->can('create_permissions');
+	Route::put('/permissions/{permission}', [PermissionController::class, 'update'])->name('permissions.update')->can('edit_permissions');
+	Route::delete('/permissions/{permission}', [PermissionController::class, 'destroy'])->name('permissions.destroy')->can('delete_permissions');
 
 	//Routes prefix
 	Route::prefix('roles')->group(function () {
 		// Add your roles routes here
-		Route::get('/', [RoleController::class, 'index'])->name('roles.index');
-		Route::post('/', [RoleController::class, 'store'])->name('roles.store');
-		Route::get('/create', [RoleController::class, 'create'])->name('roles.create');
-		Route::get('/{role}', [RoleController::class, 'show'])->name('roles.show');
-		Route::get('/{role}/edit', [RoleController::class, 'edit'])->name('roles.edit');
-		Route::put('/{role}', [RoleController::class, 'update'])->name('roles.update');
-		Route::delete('/{role}', [RoleController::class, 'destroy'])->name('roles.destroy');
+		Route::get('/', [RoleController::class, 'index'])->name('roles.index')->can('view_roles');
+		Route::post('/', [RoleController::class, 'store'])->name('roles.store')->can('create_roles');
+		Route::get('/create', [RoleController::class, 'create'])->name('roles.create')->can('create_roles');
+		Route::get('/{role}', [RoleController::class, 'show'])->name('roles.show')->can('view_roles');
+		Route::get('/{role}/edit', [RoleController::class, 'edit'])->name('roles.edit')->can('edit_roles');
+		Route::put('/{role}', [RoleController::class, 'update'])->name('roles.update')->can('edit_roles');
+		Route::delete('/{role}', [RoleController::class, 'destroy'])->name('roles.destroy')->can('delete_roles');
 		// Add your settings routes here
 	});
 
 	Route::prefix('users')->group(function () {
 		// Add your users routes here
-		Route::get('/', [UserController::class, 'index'])->name('users.index');
-		Route::post('/', [UserController::class, 'store'])->name('users.store');
-		Route::get('/create', [UserController::class, 'create'])->name('users.create');
-		Route::get('/{user}', [UserController::class, 'show'])->name('users.show');
-		Route::get('/{user}/edit', [UserController::class, 'edit'])->name('users.edit');
-		Route::put('/{user}', [UserController::class, 'update'])->name('users.update');
-		Route::delete('/{user}', [UserController::class, 'destroy'])->name('users.destroy');
+		Route::get('/', [UserController::class, 'index'])->name('users.index')->can('view_users');
+		Route::post('/', [UserController::class, 'store'])->name('users.store')->can('create_users');
+		Route::get('/create', [UserController::class, 'create'])->name('users.create')->can('create_users');
+		Route::get('/{user}', [UserController::class, 'show'])->name('users.show')->can('view_users');
+		Route::get('/{user}/edit', [UserController::class, 'edit'])->name('users.edit')->can('edit_users');
+		Route::put('/{user}', [UserController::class, 'update'])->name('users.update')->can('edit_users');
+		Route::delete('/{user}', [UserController::class, 'destroy'])->name('users.destroy')->can('delete_users');
 		// Add your settings routes here
 	});
 });
